--- next.config.mjs
@@ -1,19 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  async rewrites() {
-    return [
-      {
-        source: "/.well-known/ai-plugin.json",
-        destination: "/api/ai-plugin",
-      },
-    ];
-  },
-  webpack: (config, { isServer }) => {
-    if (isServer) {
-      config.externals = [...(config.externals || []), '@uniswap/smart-order-router'];
-    }
-    return config;
-  },
-};
-
-export default nextConfig;
