+++ package.json
@@ -2,49 +2,46 @@
   "name": "bitte-uniswap-agent",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "concurrently \"next dev --turbopack\" \"make-agent dev -p 3000\"",
-    "dev-testnet": "concurrently \"next dev --turbopack\" \"make-agent dev -p 3000 -t\"",
-    "build:next": "next build",
-    "start": "next start",
-    "lint": "next lint && prettier --check '{src/app,tests}/**/*.{js,jsx,ts,tsx}'",
-    "fmt": "next lint --fix && prettier --write '{src/app,tests}/**/*.{js,jsx,ts,tsx}'",
+    "dev": "concurrently \"tsx src/server.ts\" \"make-agent dev -p 3000\"",
+    "dev-testnet": "concurrently \"tsx src/server.ts\" \"make-agent dev -p 3000 -t\"",
+    "build": "tsc",
+    "start": "tsx --trace-deprecation src/server.ts",
+    "lint": "eslint '{src,tests}/**/*.{js,jsx,ts,tsx}'",
+    "fmt": "eslint --fix '{src,tests}/**/*.{js,jsx,ts,tsx}' && prettier --write '{src,tests}/**/*.{js,jsx,ts,tsx}'",
     "test": "jest --config jest.config.ts"
   },
   "dependencies": {
     "@bitteprotocol/agent-sdk": "^0.0.10",
     "@uniswap/sdk-core": "^6.1.0",
     "@uniswap/smart-order-router": "^4.8.6",
+    "cors": "^2.8.5",
     "ethers": "^5.7.2",
+    "express": "^4.18.2",
     "near-safe": "^0.8.8",
-    "next": "15.0.3",
-    "react": "19.0.0-rc-69d4b800-20241021",
-    "react-dom": "19.0.0-rc-69d4b800-20241021",
-    "swagger-ui-react": "^5.18.2",
-    "vercel-url": "^0.2.4",
+    "swagger-ui-express": "^5.0.0",
     "viem": "^2.21.54",
     "zerion-sdk": "^0.0.10"
   },
   "devDependencies": {
-    "@types/bun": "^1.1.14",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.10.1",
-    "@types/react": "npm:types-react@19.0.0-rc.1",
-    "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
-    "@types/swagger-ui-react": "^4.18.3",
+    "@types/swagger-ui-express": "^4.1.6",
     "concurrently": "^9.1.0",
     "dotenv": "^16.4.7",
     "eslint": "^8.57.1",
-    "eslint-config-next": "15.0.3",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "jest": "^29.7.0",
     "make-agent": "^0.0.25",
     "prettier": "^3.4.2",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
+    "ts-node-dev": "^2.0.0",
+    "tsx": "^4.19.2",
     "typescript": "^5.7.2"
-  },
-  "overrides": {
-    "@types/react": "npm:types-react@19.0.0-rc.1",
-    "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1"
   }
 }
