+++ tests/util.spec.ts
@@ -4,7 +4,7 @@ import {
   NATIVE_ASSET,
   OrderKind,
   sellTokenApprovalTx,
-} from "@/src/app/api/tools/util";
+} from "../src/tools/util";
 import { checksumAddress, getAddress, zeroAddress } from "viem";
 
 const SEPOLIA_DAI = getAddress("0xb4f1737af37711e9a5890d9510c9bb60e170cb0d");
